Increase email connection timeout to 8 seconds
Our new email provider seems to intermittently take several seconds to
establish the connection. This was causing multiple connection failures
per day with the default timeout. (Seen in SMF logs.)

This patch bumps the email connection timeout from 3 to 8 seconds.
Which, from my experimenting with this value, seems to be enough to
reliably make the connection.